Move canvas into separate module
